+++ skrl/models/torch/base.py
@@ -756,7 +756,7 @@ class Model(torch.nn.Module):
 
             # broadcast model parameter from worker/process with rank 1
             >>> if config.torch.is_distributed:
-            ...     model.update_parameters(source_model, rank=1)
+            ...     model.broadcast_parameters(rank=1)
         """
         object_list = [self.state_dict()]
         torch.distributed.broadcast_object_list(object_list, rank)