+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [1.3.0] - Unreleased
+### Added
+- Distributed multi-GPU and multi-node learning (JAX implementation)
+- Utilities to start multiple processes from a single program invocation for distributed learning using JAX
+
 ### Changed
 - Move the KL reduction from the PyTorch `KLAdaptiveLR` class to each agent using it in distributed runs
 - Move the PyTorch distributed initialization from the agent base class to the ML framework configuration